backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
playwright==1.40.0
opencv-python
mtcnn>=0.1.1
tensorflow>=2.0.0
ultralytics>=8.0.0
openai>=1.0.0

